pyproject.toml
[project]
name = "hprpg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.14",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "slack-bolt>=1.23.0",
    "starlette>=0.46.1",
    "uvicorn>=0.34.0",
    "uvloop>=0.21.0",
]

[project.scripts]
hprpg = "hprpg.__main__:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
]
